Create component file
