+++ starport/services/scaffolder/type.go
@@ -208,20 +208,6 @@ func isTypeCreated(appPath, moduleName, typeName string) (isCreated bool, err er
 	return
 }
 
-// isMsgServerDefined checks if the module uses the MsgServer convention for transactions
-// this is checked by verifying the existence of the tx.proto file
-func isMsgServerDefined(appPath, moduleName string) (bool, error) {
-	txProto, err := filepath.Abs(filepath.Join(appPath, "proto", moduleName, "tx.proto"))
-	if err != nil {
-		return false, err
-	}
-
-	if _, err := os.Stat(txProto); os.IsNotExist(err) {
-		return false, nil
-	}
-	return true, err
-}
-
 // isForbiddenTypeField returns true if the name is forbidden as a field name
 func isForbiddenTypeField(name string) bool {
 	switch name {
@@ -233,71 +219,4 @@ func isForbiddenTypeField(name string) bool {
 	}
 
 	return isGoReservedWord(name)
-}
-
-// isForbiddenComponentName returns true if the name is forbidden as a component name
-func isForbiddenComponentName(name string) bool {
-	switch name {
-	case
-		"logger",
-		"keeper",
-		"query",
-		"genesis",
-		"types",
-		"tx":
-		return true
-	}
-
-	return isGoReservedWord(name)
-}
-
-
-func isGoReservedWord(name string) bool {
-	// Check keyword or literal
-	if token.Lookup(name).IsKeyword() {
-		return true
-	}
-
-	// Check with builtin identifier
-	switch name {
-	case
-		"panic",
-		"recover",
-		"append",
-		"bool",
-		"byte",
-		"cap",
-		"close",
-		"complex",
-		"complex64",
-		"complex128",
-		"uint16",
-		"copy",
-		"false",
-		"float32",
-		"float64",
-		"imag",
-		"int",
-		"int8",
-		"int16",
-		"uint32",
-		"int32",
-		"int64",
-		"iota",
-		"len",
-		"make",
-		"new",
-		"nil",
-		"uint64",
-		"print",
-		"println",
-		"real",
-		"string",
-		"true",
-		"uint",
-		"uint8",
-		"uintptr":
-		return true
-	}
-	return false
 }
